+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/SingleStructureDataProvider.java
@@ -48,6 +48,11 @@ public class SingleStructureDataProvider implements StructureDataProvider {
     }
 
     @Override
+    public void enterWriteMode() {
+        immutable();
+    }
+
+    @Override
     public void writeRenumbered(String structureIdentifier, MmCifFile mmCifFile) {
         immutable();
     }